fix(#663): extract PollResultBuilder, remove type_body_length suppression
All pure/transform helpers from RunnerStoreState.swift extensions are
moved into a new PollResultBuilder struct in its own file. The two
public builders (buildJobState, buildGroupState) become static methods
that receive fetch/API behaviour via closure parameters, removing their
dependency on the RunnerStore singleton.

RunnerStoreState.swift now contains only thin extension wrappers that
delegate to PollResultBuilder, keeping call sites in RunnerStore.fetch()
unchanged.

The // swiftlint:disable:next type_body_length suppression is removed
from RunnerStore.swift because the type body is now well within limits.

Closes #663
